+++ packages/flutter_form_builder/lib/flutter_form_builder.dart
@@ -18,3 +18,4 @@ export 'src/fields/form_builder_switch.dart';
 export 'src/fields/form_builder_text_field.dart';
 export 'src/widgets/grouped_checkbox.dart';
 export 'src/widgets/grouped_radio.dart';
+export 'src/options/form_builder_chip_option.dart';